requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
audioread==3.1.0
backcall==0.2.0
bitsandbytes==0.49.0
bleach==6.0.0
boto3==1.42.8
botocore==1.42.8
brotli==1.2.0
cached_path==1.8.0
cachetools==6.2.2
certifi==2025.11.12
cffi==1.15.1
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.3.3
cycler==0.12.1
datasets==4.4.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
dotenv==0.9.9
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
encodec==0.1.1
executing==1.2.0
-e git+https://github.com/SWivid/F5-TTS.git@9ae46c8360303417489d2c1071f29972cd8ab171#egg=f5_tts
fastapi==0.124.4
fastjsonschema==2.16.2
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.1
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0rc3
google-genai==1.55.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gradio==5.50.0
gradio_client==1.14.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.11
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.1
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.9
langdetect==1.0.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.6.1
nltk==3.9.2
noisereduce==3.0.3
notebook==6.5.2
notebook_shim==0.2.2
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnxruntime==1.23.2
openai-whisper==20250625
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==11.3.0
piper-tts==1.3.0
platformdirs==2.6.2
pooch==1.8.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
psutil==7.1.3
pure-eval==0.2.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pypinyin==0.55.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.2.1
python-json-logger==2.0.4
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rjieba==0.1.13
rsa==4.9.1
ruff==0.14.9
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
semantic-version==2.10.0
Send2Trash==1.8.0
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.2
starlette==0.50.0
swagger-client==1.0.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.17.1
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1
torchaudio==2.9.1
torchcodec==0.9.1
torchdiffeq==0.2.5
tornado==6.2
tqdm==4.67.1
traitlets==5.8.1
transformers==4.57.3
transformers-stream-generator==0.0.5
triton==3.5.1
typer==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
uri-template==1.2.0
urllib3==2.6.2
uvicorn==0.38.0
vocos==0.1.0
wandb==0.23.1
wcwidth==0.2.6
webcolors==1.12
websocket-client==1.5.0
websockets==15.0.1
widgetsnbextension==4.0.5
x-transformers==2.11.23
xxhash==3.6.0
yarl==1.22.0
